reduce duplication between elm & javascript changed dependencies
