+++ lib/deploy_complexity/changed_elm_packages.rb
@@ -1,22 +1,10 @@
 # frozen_string_literal: true
 
 require 'json'
+require 'deploy_complexity/changed_dependencies'
 
 # Takes in two elm.json files and detects which packages have changed
-class ChangedElmPackages
-  def initialize(old:, new:)
-    @old_dependencies = all_dependencies(JSON.parse(old))
-    @new_dependencies = all_dependencies(JSON.parse(new))
-  end
-
-  def changes
-    [
-      format_dependencies("Added", added_dependencies),
-      format_dependencies("Removed", removed_dependencies),
-      format_updated_dependencies(updated_dependencies)
-    ].flatten
-  end
-
+class ChangedElmPackages < ChangedDependencies
   private
 
   def all_dependencies(json)
@@ -29,37 +17,5 @@ class ChangedElmPackages
       test_dependencies.fetch("direct"),
       test_dependencies.fetch("indirect")
     ].inject(&:merge)
-  end
-
-  def added_dependencies
-    @new_dependencies.dup.delete_if { |package, _| @old_dependencies.key?(package) }
-  end
-
-  def removed_dependencies
-    @old_dependencies.dup.delete_if { |package, _| @new_dependencies.key?(package) }
-  end
-
-  def updated_dependencies
-    @new_dependencies.each_with_object({}) do |(package, new_version), changed_dependencies|
-      next if @old_dependencies[package].nil?
-      next if @old_dependencies[package] == new_version
-
-      changed_dependencies[package] = {
-        old: @old_dependencies[package],
-        new: new_version
-      }
-    end
-  end
-
-  def format_dependencies(label, dependencies)
-    dependencies.map do |(package, version)|
-      "#{label} #{package}: #{version}"
-    end
-  end
-
-  def format_updated_dependencies(dependencies)
-    dependencies.map do |(package, versions)|
-      "Updated #{package}: #{versions.fetch(:old)} -> #{versions.fetch(:new)}"
-    end
   end
 end
